Ran experiments for Watershed. JMeter was defaulting to header-> Content-Type: "application/x-www-form-urlencoded". This went unnoticed while working with LearningLocker. However, with Watershed, this resulted in POST requests being handled as GET, with no error messages produced. Normal functionality was restored by adding the header parameter Content-Type: "application/json".
Further, the presence of an authority field in a number of json resulted in a 403, Forbidden error. Deleting the field resolved the issue.

A note on the file namings: "examplesPavlos_LearningLocker.jmx" and "examplesPavlos_Watershed.jmx" contain the final states of the executed recipes for the set of experiments involving the respective LRS. "examplePavlos.jmx" is the current state of experiments. Other ".jmx" highlight specific situations as described in previous commits (as the filename indicates).